Flush logs before each example
It is possible to have logging after the `after` hooks, which then leak
into the following example. Wiping the logs before the example prevents
this leakage.

This is the cause of a few flaky specs in Up. Namely, all the ones in
`api/spec/end_to_end/admin/account_owner_risk_spec.rb`.